requirements.txt
streamlit>=1.22.0
boto3>=1.26.0
python-dotenv>=1.0.0
pillow>=9.5.0
plotly>=5.14.0
pandas>=1.5.3 